+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-* 
+* FloorMap - Improve configuration error alerting.
 * FloorMap - Fix bug in handling config provided via settings vice js.
 * FloorMap - Make zone editing configurable.
 * FloorMap/GeoMap - use small circle markers if icon not set in data.